docker: cross compile binary for Dockerfile
This cross compiles binary before COPY it to the Docker image.My machine
was heating up and taking forever on the

```
RUN CGO_ENABLED=0 GOOS=linux GOARCH=amd64 GO111MODULE=on go build -a -o manager main.go
```

instruction in Dockerfile. There is no need to build the binary inside the image because
go already supports cross compilation. In this change we build manager.exe locally using the
vendor modules and copy it into the image.

This adds `cross.Dockerfile` which is the Dockerfile with instruction to use the locally built
manager binary.

This is only for testing purposes because we will be building the images
very often and any time saved is time well spent.

calling `make cross-build-image` will now build `tyk-operator:latest` .
Please use this in development.