add tools package
This is a commandline utility to provide helpers for working with the
operator.

In this commit tools migrate command is added with initial implementation
of migrating from dashboard to the operator resources

Things that are done so far

- APIDefinition to ApiDefinition resource conversion
- Policy to SecurityPolicy conversion

This command calls the dashboard for all api's and policies and generates
yaml resource files using the data. The result is written on stdout.